// src/lib/types/tp.d.ts
export type WorkoutTP = {
  AthleteId?: number;
  CadenceAverage?: number;
  CadenceMaximum?: number;
  Calories?: number;
  CaloriesPlanned?: number;
  Completed?: boolean;
  Distance?: number;
  ElevationAverage?: number;
  ElevationGain?: number;
  ElevationLoss?: number;
  ElevationMaximum?: number;
  ElevationMinimum?: number;
  Energy?: number;
  HeartRateAverage?: number;
  HeartRateMaximum?: number;
  HeartRateMinimum?: number;
  IF?: number;
  IFPlanned?: number;
  Id: number;
  NormalizedPower?: number;
  PowerAverage?: number;
  PowerMaximum?: number;
  StartTime?: Date;
  StartTimePlanned?: Date;
  TempAvg?: number;
  TempMax?: number;
  TempMin?: number;
  TotalTime?: number;
  TotalTimePlanned?: number;
  TssActual?: number;
  TssCalculationMethod?: string;
  VelocityAverage?: number;
  VelocityMaximum?: number;
  WorkoutDay?: Date;
  WorkoutType?: string;
  Description?: string;
  Title?: string;
};